+++ src/lib/events.js
@@ -3,3 +3,13 @@ class Events {}
  * Event fired when a new {Tweet} was created.
  */
 Events.NEW_TWEET = 'new_tweet';
+
+/**
+ * Event fired when scroll to last read starts.
+ */
+Events.SCROLL_TO_LAST_READ_START = 'scroll_to_last_read_start';
+
+/**
+ * Event fired when scroll to last read stops.
+ */
+Events.SCROLL_TO_LAST_READ_STOP = 'scroll_to_last_read_stop';