+++ src/util/stringbigint.js
@@ -3,18 +3,18 @@
  * This class will convert integers to zero padded strings.
  * Comparison is then simply done as string comparison.
  */
-class BigInt {
+class StringBigInt {
 	/**
-	 * @param {int} value 
+	 * @param {int} value
 	 */
 	constructor (value) {
 		// Number of 'digits' to use so we can always do string comparison of tweet ids by prepending zeros.
-		this.valueSize = 24; 
-		this.value = new String(value);
+		this.valueSize = 24;
+		this.value = this.padZeros('' + value);
 	}
 
 	/**
-	 * @param {int} val 
+	 * @param {int} val
 	 */
 	padZeros(val) {
 		while (val.length < this.valueSize) {
@@ -41,6 +41,16 @@ class BigInt {
 	 * @return {string}
 	 */
 	valueOf() {
-		return this.padZeros(this.value);
+		return this.value;
+	}
+
+	/**
+	 * Compare with other value.
+	 * @param {StringBigInt} other
+	 * @return int From string comparison.
+	 */
+	compare(other) {
+		deb.debug('StringBigInt::compare', this.toString(), other.toString(), this.toString().localeCompare(other.toString(), 'en'));
+		return this.toString().localeCompare(other.toString(), 'en');
 	}
 }